Remove unused DOMAttrModified from EventsNames and Document::ListenerType https://bugs.webkit.org/show_bug.cgi?id=97591
Reviewed by Ojan Vafai.

WebKit does not, and will never, fire DOMAttrModified events, so
there's no need to create the "DOMAttrModified" event name.

The only use of the name was to set the DOMATTRMODIFIED_LISTENER bit
on Document; with the name gone, the enum value can be removed as well.

* dom/Document.cpp:
(WebCore::Document::addListenerTypeIfNeeded):
* dom/Document.h: Remove DOMATTRMODIFIED_LISTENER, and switch this
enum to use shift-left instead of hex values, so as to be easier to
update in future.
* dom/EventNames.h:
(WebCore):


Identifier: 115569@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129551 268f45cc-cd09-0410-ab3c-d52691b4dbfc