webkit-patch rebaseline-expectations is broken https://bugs.webkit.org/show_bug.cgi?id=97619
Reviewed by Ojan Vafai.

The new expectations format is causing two problems for
webkit-patch rebaseline expectations

1) the transformation into and out of the old syntax (which is
still used internally) is somewhat lossy, e.g., we're not
preserving the case of Bug(x) identifiers. Also, we can't
tell if the input was [ WontFix ] or [ Skip WontFix Pass ]
2) the new syntax is more lenient, allowing for multiple ways to
specify the same result, e.g., "[ Skip ]" may or may not be
missing.

As a result, we end up rewriting a whole bunch of lines for
purely syntactic differences.

Both of these things are bugs that need to be fixed, but
we should also not just rewrite lines that shouldn't be
modified.

This patch fixes the last issue (no rewriting).

* Scripts/webkitpy/layout_tests/models/test_expectations.py:
(TestExpectationLine._format_line):
(TestExpectations.remove_rebaselined_tests):
* Scripts/webkitpy/layout_tests/models/test_expectations_unittest.py:
(RebaseliningTest.test_remove):

Identifier: 115946@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129941 268f45cc-cd09-0410-ab3c-d52691b4dbfc