[EFL] Unskip {horizontal,vertical}-scrollbar-in-rtl.html
The expectations in fast/dom are actually Mac-specific (cf.
r73063), and most other ports use their own platform-specific
results due to the differences in behavior of the Home/End keys.

The results being checked in here are in line with the other
ports, so everything seems to be OK.

* platform/efl/TestExpectations:
* platform/efl/fast/dom/horizontal-scrollbar-in-rtl-expected.txt: Added.
* platform/efl/fast/dom/vertical-scrollbar-in-rtl-expected.txt: Added.


Identifier: 115990@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@129986 268f45cc-cd09-0410-ab3c-d52691b4dbfc