webkitpy should accept a different httpd.conf specified by the user https://bugs.webkit.org/show_bug.cgi?id=98071
Reviewed by Dirk Pranke.

The existing httpd.conf variants (or lighttpd.conf, for that
matter) we have do not always suit the user's system. This is
particularly true on Linux/Unix, where Apache can be installed in
a plethora of ways and the LoadModule calls can fail to specify
the proper module paths.

For now, we start accepting the WEBKIT_HTTP_SERVER_CONF_PATH
environment variable, which allows the user to specify the
absolute path to another http server configuration file that might
work on the user's system.

In the long term, we should try to generate our configuration file
and stop requiring all the different httpd.conf files we have as
well as this hack.

* Scripts/webkitpy/layout_tests/port/base.py:
(Port._path_to_apache_config_file):
* Scripts/webkitpy/layout_tests/port/port_testcase.py:
(test_path_to_apache_config_file):


Identifier: 116084@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130083 268f45cc-cd09-0410-ab3c-d52691b4dbfc