Remove cases of testRunner.disableImageLoading() from the layout tests. https://bugs.webkit.org/show_bug.cgi?id=98187
Reviewed by Simon Fraser.

According to the discussion started in [1], it does not make much
sense to keep disableLoadingImages() around when
overridePreferences("WebKitDisplayImagesKey", 0) achieves the same
effect.

There were only two cases of disableLoadingImages() in the tests;
one of the tests is disabled and the other is currently skipped by
all ports, so there should be no regressions, such as problems
with ports that may implement testRunner.disableLoadingImages()
but not overridePreferences("WebKitDisplayImagesKey").

[1] http://lists.webkit.org/pipermail/webkit-dev/2012-September/022323.html

* http/tests/misc/favicon-loads-with-icon-loading-override.html:
* http/tests/misc/favicon-loads-with-images-disabled.html:


Identifier: 116200@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130201 268f45cc-cd09-0410-ab3c-d52691b4dbfc