Implement LocaleICU::dateFormat https://bugs.webkit.org/show_bug.cgi?id=98118
Reviewed by Hajime Morita.

http://trac.webkit.org/changeset/130127 introduced
Localizer::dateFormat, and this is its implementation for LocaleICU
classs. The code is going to be used when
ENABLE_INPUT_MULTIPLE_FIELDS_UI is enabled.

No new tests. The function is not used yet.

* platform/text/LocaleICU.cpp:
(WebCore::LocaleICU::dateFormat):
Implemented. Note that m_shortDateFormat is a UDateFormat object, which
knows various format information.
* platform/text/LocaleICU.h:
(LocaleICU): Add m_dateFormat to cache the format string.


Identifier: 116242@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130244 268f45cc-cd09-0410-ab3c-d52691b4dbfc