Exception thrown when running accessibility/container-node-delete-causes-crash.html test https://bugs.webkit.org/show_bug.cgi?id=98291
Reviewed by Andreas Kling.

The accessibility/container-node-delete-causes-crash.html test will cause a full accessibility tree
to be created by trying to look up an element with a non-existent ID. This caused an exception to be thrown
when trying to access the children of an element that didn't have any children. Fix this by adding
BEGIN_AX_OBJC_EXCEPTIONS/END_AX_OBJC_EXCEPTIONS around the call to get the children.

* DumpRenderTree/mac/AccessibilityControllerMac.mm:
(findAccessibleObjectById):

Identifier: 116309@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130312 268f45cc-cd09-0410-ab3c-d52691b4dbfc