Add a performance test for subtree detaching https://bugs.webkit.org/show_bug.cgi?id=98281
Reviewed by Eric Seidel.

The new test tries to stress detaching on a heavy subtree (2 levels deep but with 1,000 nodes per level).
We set display: none on the root and measure the time it takes to update the tree. Note that we will also
measure the time taken by recalcStyle and relayout but I don't think we can avoid that.

* Layout/subtree-detaching.html: Added.


Identifier: 116338@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130341 268f45cc-cd09-0410-ab3c-d52691b4dbfc