Crash in Safari at com.apple.JavaScriptCore: WTF::StringImpl::is8Bit const + 12 https://bugs.webkit.org/show_bug.cgi?id=98433
Reviewed by Jessie Berlin.

The problem is due to a String with a null StringImpl (i.e. a null string).
Added a length check before the is8Bit() check since length() checks for a null StringImpl.  Changed the
characters16() call to characters() since it can handle a null StringImpl as well.

* API/JSValueRef.cpp:
(JSValueMakeFromJSONString):


Identifier: 116420@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130424 268f45cc-cd09-0410-ab3c-d52691b4dbfc