Font::glyphDataAndPageForCharacter doesn't account for text orientation when using systemFallback on a cold cache. https://bugs.webkit.org/show_bug.cgi?id=98452.
Reviewed by Dan Bernstein.

Source/WebCore:

The text orientation was considered only when there is a cache hit.
This change moves the logic to handle text orientation to a separate
inline function that is called also when the glyph is added to the cache.

Test: fast/text/vertical-rl-rtl-linebreak.html

* platform/graphics/FontFastPath.cpp:
(WebCore::applyTextOrientationForCharacter): Added.
(WebCore::Font::glyphDataAndPageForCharacter): Modified to use the new function in
both cases of cold and warm cache.

LayoutTests:

The new test adds platform specific results for Mac, since it uses system fallback fonts.
I've added the expected failure for all the other platforms.

* fast/text/vertical-rl-rtl-linebreak.html: Added.
* platform/chromium/TestExpectations:
* platform/efl/TestExpectations:
* platform/gtk/TestExpectations:
* platform/mac/fast/text/vertical-rl-rtl-linebreak-expected.txt: Added.
* platform/qt/TestExpectations:
* platform/win/TestExpectations:



Identifier: 116438@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130443 268f45cc-cd09-0410-ab3c-d52691b4dbfc