[BlackBerry] Fix regression in proxy auth https://bugs.webkit.org/show_bug.cgi?id=98533
Patch by Joe Mason <jmason@rim.com> on 2012-10-05
Reviewed by Yong Li.

The proxy auth dialog gets the proxy address from
BlackBerry::Platform::Settings::proxyAddress, which returns
"host:port", but we try to parse it with a KURL, which expects
"scheme://host:port".  Since this is an http proxy, add http:// to the
url to get it to parse.

PR 220567.

* platform/network/blackberry/NetworkJob.cpp:
(WebCore::NetworkJob::sendRequestWithCredentials):

Identifier: 116527@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130537 268f45cc-cd09-0410-ab3c-d52691b4dbfc