[GTK] Test /webkit2/WebKitWebView/populate-menu asserts in debug bots https://bugs.webkit.org/show_bug.cgi?id=98507
Reviewed by Martin Robinson.

Make sure we retain the page URL before asking for the icon.

* UIProcess/API/gtk/WebKitFaviconDatabase.cpp:
(getIconSurfaceSynchronously): Retain the page URL before asking
for the icon, and release it if we fail to get an icon.
(processPendingIconsForPageURL): Renamed to make it clear it
receives a Sring and not a char* and moved
getIconSurfaceSynchronously() call out of the loop to call it only
once even if there are multiple requests pending.
(iconDataReadyForPageURLCallback): Do not retain the page URL
here.

Identifier: 116612@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130625 268f45cc-cd09-0410-ab3c-d52691b4dbfc